refactor: breakdown knowledge base hooks and components into feature-specific folders
