+++ apps/web/src/features/locations/hooks/mutations/use-delete-location.ts
