+++ apps/web/src/features/locations/hooks/queries/use-location-list.ts
