+++ apps/web/src/features/props/hooks/mutations/use-delete-prop.ts
