+++ apps/web/src/features/wildcards/hooks/mutations/use-delete-wildcard.ts
