+++ apps/web/src/features/wildcards/hooks/queries/use-wildcard-list.ts
